# matchSet.addX / matchSet.addXSingleThreaded performance improvements