- addX: adding length check to avoid copying the map if nothing was changed.
- Introducing addXSingleThreaded which allows update-in-place in single-threaded usages.

```
name \ time/op                          benchs/1_baseline  benchs/2_len_check  benchs/3_thread_safe_add  benchs/4_thread_safe_len_check
_Quamina-10                                   48.0µs ± 0%         33.9µs ± 0%                6.6µs ± 0%                      6.8µs ± 0%

name \ alloc/op                         benchs/1_baseline  benchs/2_len_check  benchs/3_thread_safe_add  benchs/4_thread_safe_len_check
_Quamina-10                                   41.0kB ± 0%         30.4kB ± 0%                8.9kB ± 0%                      8.9kB ± 0%

name \ allocs/op                        benchs/1_baseline  benchs/2_len_check  benchs/3_thread_safe_add  benchs/4_thread_safe_len_check
_Quamina-10                                      161 ± 0%            129 ± 0%                   28 ± 0%                         28 ± 0%
```

* `benchs/1_baseline` - baseline
* `benchs/2_len_check` - adding early returrn with length check to `addX` (still being used in matching)
* `benchs/3_thread_safe_add` - using `addXSingleThreaded` in matching
* `benchs/4_thread_safe_len_check` - adding early return with length check to `addXSingleThreaded` .. didn't improved the situation, which makes sense in this case.