@yosiat 2022-09-27
There are two flows which aren't covered:
1. `addX` called with empty exes
2. `addX` called with existing set and we are adding exes

This can come up from the flow of exists false case, If I would add another "exists false" example to this test - `TestExerciseMatching` it will take care of number 2.

I see there are no direct unit tests for `matchSet`, I can add direct tests and this will fix up the coverage issue.
Is that desired? or we want to come-up with new tests for "exists false" that will cover this case?