@timbray 2022-09-27
> I see there are no direct unit tests for `matchSet`, I can add direct tests and this will fix up the coverage issue. Is that desired? or we want to come-up with new tests for "exists false" that will cover this case?

Up to you, I'm sure you'll do the right thing.  I just hate to see any check-in which reduces coverage.